+++ src/client/components/Controllers/Base/Panels/Report.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react'
 import Box from '@mui/material/Box'
-import Report from '../../../Editors/Report'
+import Report from '../../../Modules/Report'
 import * as types from '../../../../types'
 
 export interface ReportPanelProps {
